+++ packages/netlify-cms-core/src/backend.js
@@ -42,7 +42,9 @@ const slugFormatter = (collection, entryData, slugConfig) => {
 
   const identifier = entryData.get(selectIdentifier(collection));
   if (!identifier) {
-    throw new Error('Collection must have a field name that is a valid entry identifier');
+    throw new Error(
+      'Collection must have a field name that is a valid entry identifier, or must have `identifier_field` set',
+    );
   }
 
   const slug = template
@@ -68,6 +70,9 @@ const slugFormatter = (collection, entryData, slugConfig) => {
     })
     // Convert slug to lower-case
     .toLocaleLowerCase()
+
+    // Remove single quotes.
+    .replace(/[']/g, '')
 
     // Replace periods with dashes.
     .replace(/[.]/g, '-');
@@ -95,7 +100,7 @@ const commitMessageFormatter = (type, config, { slug, path, collection }) => {
       case 'path':
         return path;
       case 'collection':
-        return collection.get('label');
+        return collection.get('label_singular') || collection.get('label');
       default:
         console.warn(`Ignoring unknown variable “${variable}” in commit message template.`);
         return '';
@@ -444,7 +449,7 @@ class Backend {
     }
 
     const commitMessage = commitMessageFormatter('delete', config, { collection, slug, path });
-    return this.implementation.deleteFile(path, commitMessage);
+    return this.implementation.deleteFile(path, commitMessage, { collection, slug });
   }
 
   deleteMedia(config, path) {
@@ -495,9 +500,13 @@ class Backend {
   }
 
   filterEntries(collection, filterRule) {
-    return collection.entries.filter(
-      entry => entry.data[filterRule.get('field')] === filterRule.get('value'),
-    );
+    return collection.entries.filter(entry => {
+      const fieldValue = entry.data[filterRule.get('field')];
+      if (Array.isArray(fieldValue)) {
+        return fieldValue.includes(filterRule.get('value'));
+      }
+      return fieldValue === filterRule.get('value');
+    });
   }
 }