+++ website/gatsby-config.js
@@ -12,12 +12,16 @@ module.exports = {
           title: 'Quick Start',
         },
         {
-          name: 'guides',
-          title: 'Guides',
-        },
-        {
           name: 'reference',
           title: 'Reference',
+        },
+        {
+          name: 'media',
+          title: 'Media',
+        },
+        {
+          name: 'guides',
+          title: 'Guides',
         },
         {
           name: 'contributing',
@@ -34,8 +38,6 @@ module.exports = {
         name: 'content',
       },
     },
-    'gatsby-transformer-yaml',
-    'gatsby-transformer-json',
     {
       resolve: 'gatsby-source-filesystem',
       options: {
@@ -49,13 +51,26 @@ module.exports = {
         // prettier-ignore
         plugins: [
           'gatsby-remark-autolink-headers',
-          'gatsby-remark-prismjs'
+          {
+            resolve: 'gatsby-remark-prismjs',
+            options: {
+              noInlineHighlight: true,
+            },
+          },
         ]
       },
     },
+    'gatsby-transformer-yaml',
+    'gatsby-transformer-json',
     'gatsby-plugin-postcss',
     'gatsby-plugin-react-helmet',
     'gatsby-plugin-catch-links',
+    {
+      resolve: 'gatsby-plugin-netlify-cms',
+      options: {
+        modulePath: `${__dirname}/src/cms/cms.js`,
+      },
+    },
     {
       resolve: `gatsby-plugin-manifest`,
       options: {