+++ packages/netlify-cms/CHANGELOG.md
@@ -3,6 +3,84 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.2](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.3.1...netlify-cms@2.3.2) (2018-12-27)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+## [2.3.1](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.3.0...netlify-cms@2.3.1) (2018-12-11)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+# [2.3.0](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.2.1...netlify-cms@2.3.0) (2018-12-04)
+
+
+### Features
+
+* add cloudinary support ([#1932](https://github.com/netlify/netlify-cms/issues/1932)) ([1fc2f50](https://github.com/netlify/netlify-cms/commit/1fc2f50))
+
+
+
+
+
+## [2.2.1](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.2.0...netlify-cms@2.2.1) (2018-11-29)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+# [2.2.0](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.1.3...netlify-cms@2.2.0) (2018-11-12)
+
+**Note:** Version bump only for package netlify-cms
+
+
+
+
+
+<a name="2.1.3"></a>
+## [2.1.3](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.1.2...netlify-cms@2.1.3) (2018-10-09)
+
+
+
+
+**Note:** Version bump only for package netlify-cms
+
+<a name="2.1.2"></a>
+## [2.1.2](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.1.1...netlify-cms@2.1.2) (2018-09-17)
+
+
+
+
+**Note:** Version bump only for package netlify-cms
+
+<a name="2.1.1"></a>
+## [2.1.1](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.1.0...netlify-cms@2.1.1) (2018-09-06)
+
+
+
+
+**Note:** Version bump only for package netlify-cms
+
+<a name="2.1.0"></a>
+# [2.1.0](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.0.11...netlify-cms@2.1.0) (2018-09-06)
+
+
+### Features
+
+* **media:** add external media library support, Uploadcare integration ([#1602](https://github.com/netlify/netlify-cms/issues/1602)) ([0596904](https://github.com/netlify/netlify-cms/commit/0596904))
+
+
+
+
 <a name="2.0.11"></a>
 ## [2.0.11](https://github.com/netlify/netlify-cms/compare/netlify-cms@2.0.10...netlify-cms@2.0.11) (2018-08-27)